+++ util-stats/src/test/java/com/twitter/compiletest/MetricBuilderCompilationTest.java
@@ -4,14 +4,11 @@ import org.junit.Test;
 
 import scala.Some;
 
-import com.twitter.finagle.stats.Counter;
-import com.twitter.finagle.stats.Gauge;
 import com.twitter.finagle.stats.InMemoryStatsReceiver;
 import com.twitter.finagle.stats.MetricBuilder;
 import com.twitter.finagle.stats.MetricTypes;
 import com.twitter.finagle.stats.Microseconds;
 import com.twitter.finagle.stats.NoRoleSpecified;
-import com.twitter.finagle.stats.Stat;
 import com.twitter.finagle.stats.StatsReceiver;
 import com.twitter.finagle.stats.Verbosity;
 
@@ -44,17 +41,5 @@ public final class MetricBuilderCompilationTest {
       .withName("my", "very", "cool", "name")
       .withRelativeName("cool", "name")
       .withPercentiles(0.99, 0.88, 0.77);
-  }
-
-  @Test
-  public void testConstructingMetrics() {
-    StatsReceiver sr = new InMemoryStatsReceiver();
-    MetricBuilder gmb = sr.metricBuilder(MetricTypes.GAUGE_TYPE);
-    MetricBuilder smb = sr.metricBuilder(MetricTypes.HISTOGRAM_TYPE);
-    MetricBuilder cmb = sr.metricBuilder(MetricTypes.COUNTER_TYPE);
-
-    Gauge g = gmb.gauge(() -> 3.0f, "my", "cool", "gauge");
-    Stat s = smb.histogram("my", "cool", "histogram");
-    Counter c = cmb.counter("my", "cool", "counter");
   }
 }